use correct types

+++ firedrake/evaluate.h
@@ -54,7 +54,7 @@ extern int locate_cell_from_candidates(struct Function *f,
 		       void *found_ref_coords,
 		       double *found_ref_cell_dist_l1,
 			   size_t nids,
-			   size_t *ids,
+			   int64_t *ids,
 			   size_t ncells_ignore,
 			   int* cells_ignore);
